requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.8,<1.0
gunicorn>=22.0,<24.0
requests>=2.31,<3.0
firebase-admin>=6.4,<7.0
